+++ python/jupytergis_core/jupytergis_core/processing.py
@@ -1,15 +1,9 @@
-import json
 import logging
 import os
 import shutil
 import subprocess
 import tempfile
 from base64 import b64encode
-from typing import Any
-
-import tornado
-from jupyter_server.base.handlers import APIHandler
-from jupyter_server.utils import url_path_join
 
 logger = logging.getLogger(__name__)
 
@@ -17,126 +11,12 @@ logger = logging.getLogger(__name__)
 ALLOWED_OPERATIONS = {"ogr2ogr", "gdal_rasterize", "gdalwarp", "gdal_translate"}
 
 
-def _gdal_available() -> bool:
+def gdal_available() -> bool:
     """Check if GDAL CLI tools are available on the system."""
     return shutil.which("ogr2ogr") is not None
 
 
-class ProcessingHandler(APIHandler):
-    """Handler for server-side GDAL processing operations.
-
-    GET  — returns whether server-side GDAL is available.
-    POST — runs a GDAL CLI operation and returns the result.
-    """
-
-    @tornado.web.authenticated
-    async def get(self):
-        """Return GDAL availability status."""
-        available = _gdal_available()
-        version = None
-        if available:
-            try:
-                result = subprocess.run(
-                    ["ogr2ogr", "--version"],
-                    capture_output=True,
-                    text=True,
-                    timeout=5,
-                )
-                version = result.stdout.strip()
-            except Exception:
-                pass
-
-        self.finish(json.dumps({"available": available, "version": version}))
-
-    @tornado.web.authenticated
-    async def post(self):
-        """Run a GDAL processing operation.
-
-        Expected JSON body:
-        {
-            "operation": "ogr2ogr" | "gdal_rasterize" | ...,
-            "options": ["-f", "GeoJSON", "-dialect", "SQLITE", "-sql", "...", "output.geojson"],
-            "geojson": "<geojson string>",
-            "outputName": "output.geojson"
-        }
-
-        Returns:
-        {
-            "result": "<output content>",
-            "format": "text" | "base64"
-        }
-        """
-        try:
-            body = json.loads(self.request.body)
-        except json.JSONDecodeError:
-            self.set_status(400)
-            self.finish(json.dumps({"error": "Invalid JSON body"}))
-            return
-
-        operation = body.get("operation")
-        options = body.get("options", [])
-        geojson = body.get("geojson")
-        output_name = body.get("outputName", "output.geojson")
-
-        if operation not in ALLOWED_OPERATIONS:
-            self.set_status(400)
-            self.finish(
-                json.dumps(
-                    {
-                        "error": f"Unsupported operation: {operation}. "
-                        f"Allowed: {sorted(ALLOWED_OPERATIONS)}"
-                    }
-                )
-            )
-            return
-
-        if not geojson:
-            self.set_status(400)
-            self.finish(json.dumps({"error": "Missing 'geojson' field"}))
-            return
-
-        if not _gdal_available():
-            self.set_status(503)
-            self.finish(
-                json.dumps({"error": "GDAL CLI tools are not installed on the server"})
-            )
-            return
-
-        # Validate options — must all be strings, no shell injection
-        if not isinstance(options, list) or not all(
-            isinstance(o, str) for o in options
-        ):
-            self.set_status(400)
-            self.finish(json.dumps({"error": "'options' must be a list of strings"}))
-            return
-
-        try:
-            (
-                result_content,
-                result_format,
-            ) = await tornado.ioloop.IOLoop.current().run_in_executor(
-                None,
-                lambda: _run_gdal(operation, options, geojson, output_name),
-            )
-        except subprocess.TimeoutExpired:
-            self.set_status(504)
-            self.finish(json.dumps({"error": "GDAL operation timed out"}))
-            return
-        except subprocess.CalledProcessError as e:
-            logger.error("GDAL %s failed: %s", operation, e.stderr)
-            self.set_status(500)
-            self.finish(json.dumps({"error": f"GDAL error: {e.stderr.strip()}"}))
-            return
-        except Exception as e:
-            logger.error("Processing error: %s", e)
-            self.set_status(500)
-            self.finish(json.dumps({"error": str(e)}))
-            return
-
-        self.finish(json.dumps({"result": result_content, "format": result_format}))
-
-
-def _run_gdal(
+def run_gdal(
     operation: str,
     options: list[str],
     geojson: str,
@@ -147,7 +27,7 @@ def _run_gdal(
     Returns (content, format) where format is "text" or "base64".
     """
     with tempfile.TemporaryDirectory() as tmpdir:
-        input_path = os.path.join(tmpdir, "input.geojson")
+        input_path = os.path.join(tmpdir, "data.geojson")
         output_path = os.path.join(tmpdir, output_name)
 
         with open(input_path, "w") as f:
@@ -207,15 +87,3 @@ def _run_gdal(
         else:
             with open(output_path, "r") as f:
                 return f.read(), "text"
-
-
-def setup_processing_handlers(web_app: Any) -> None:
-    """Register the processing handler."""
-    host_pattern = ".*$"
-    base_url = web_app.settings["base_url"]
-
-    processing_route = url_path_join(base_url, "jupytergis_core", "processing")
-    handlers = [(processing_route, ProcessingHandler)]
-
-    web_app.add_handlers(host_pattern, handlers)
-    logger.info("JupyterGIS processing endpoint initialized at: %s", processing_route)
